myproject/requirements.txt
djangorestframework==3.12.0